LIT/requirements.txt
amqp==5.1.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.3
cffi==1.15.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.4
django-redis==5.3.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.2
packaging==23.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
rpds-py==0.10.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
